Shared constructor between platforms
