+++ homeassistant/components/tailscale/sensor.py
@@ -15,7 +15,6 @@ from homeassistant.components.sensor import (
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
-from homeassistant.helpers.update_coordinator import DataUpdateCoordinator
 
 from . import TailscaleEntity
 from .const import DOMAIN
@@ -67,20 +66,6 @@ class TailscaleSensorEntity(TailscaleEntity, SensorEntity):
     """Defines a Tailscale sensor."""
 
     entity_description: TailscaleSensorEntityDescription
-
-    def __init__(
-        self,
-        *,
-        coordinator: DataUpdateCoordinator,
-        device: TailscaleDevice,
-        description: TailscaleSensorEntityDescription,
-    ) -> None:
-        """Initialize a Tailscale sensor."""
-        super().__init__(coordinator=coordinator)
-        self.entity_description = description
-        self.device_id = device.device_id
-        self._attr_name = f"{device.hostname} {description.name}"
-        self._attr_unique_id = f"{device.device_id}_{description.key}"
 
     @property
     def native_value(self) -> datetime | None:
